feeder/dags_requirements.txt
beautifulsoup4>=4.14.2
langchain[google-genai]>=1.0.2
langchain-google-genai>=3.0.0
qdrant-client>=1.15.1
langchain-qdrant>=1.1.0
playwright>=1.55.0
apache-airflow==3.1.0
langchain-community>=0.4.1
langgraph>=1.0.1